# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Booking API documentation",
    "description": "L5 Swagger OpenApi description",
    "contact": {
      "email": "khujand.nanocoding@gmail.com"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:3001/api",
      "description": "Local API Server"
    }
  ],
  "paths": {
...
    "/user": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Display all users",
        "operationId": "usersAll",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterModel"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/user/{id}": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Show user by id",
        "operationId": "showUserById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The user`s id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterModel"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "User"
        ],
        "summary": "Update user by id",
        "operationId": "updateUserById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The user`s id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterStoreModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterModel"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "User"
        ],
        "summary": "Delete user by id",
        "operationId": "deleteUserById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The user`s id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Success"
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/user/toggle/{id}": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Toggle status user by id",
        "operationId": "toggleStatusUserById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The user`s id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Success"
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "RegisterModel": {
        "title": "Register model",
        "description": "Register model",
        "properties": {
          "id": {
            "title": "Id",
            "description": "Users`s id",
            "format": "string"
          },
          "name": {
            "title": "Name",
            "description": "Users`s name",
            "format": "string"
          },
          "email": {
            "title": "Email",
            "description": "Users`s email",
            "format": "email"
          },
          "password": {
            "title": "Password",
            "description": "Users`s password",
            "format": "string"
          },
          "role": {
            "title": "Role",
            "description": "Users`s role",
            "format": "string",
            "default": "user",
            "enum": [
              "admin",
              "user"
            ]
          },
          "phone": {
            "title": "Phone",
            "description": "Users`s phone",
            "format": "string",
            "example": "+992927777777"
          },
          "image": {
            "title": "Image",
            "description": "Users`s image",
            "format": "string",
            "example": "no-image.png"
          },
          "company_id": {
            "title": "Company id",
            "description": "Users`s company id",
            "format": "string"
          },
          "status": {
            "title": "Status",
            "description": "Users`s status",
            "format": "string"
          },
          "created_at": {
            "title": "Created at",
            "description": "Users`s created at",
            "format": "date-time"
          },
          "updated_at": {
            "title": "Updated at",
            "description": "Users`s updated at",
            "format": "date-time"
          }
        },
        "type": "object"
      },
      "RegisterStoreModel": {
        "title": "Register store model",
        "description": "Register store model",
        "properties": {
          "name": {
            "title": "Name",
            "description": "Users`s name",
            "format": "string"
          },
          "email": {
            "title": "Email",
            "description": "Users`s email",
            "format": "email"
          },
          "password": {
            "title": "Password",
            "description": "Users`s password",
            "format": "string"
          },
          "role": {
            "title": "Role",
            "description": "Users`s role",
            "format": "string",
            "default": "user",
            "enum": [
              "admin",
              "user"
            ]
          },
          "phone": {
            "title": "Phone",
            "description": "Users`s phone",
            "format": "string",
            "example": "+992927777777"
          },
          "image": {
            "title": "Image",
            "description": "Users`s image",
            "format": "string",
            "example": "no-image.png"
          },
          "company_id": {
            "title": "Company id",
            "description": "Users`s company id",
            "format": "string"
          },
          "status": {
            "title": "Status",
            "description": "Users`s status",
            "format": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}